ca0132_dsp_functions: Add op length suffix handling.
Add the ability to signify the op_length that is desired by adding the
length as a suffix on the end of the assembly op, i.e 'MOV:4'. This
helps the assembler select the exact op. If the length isn't defined,
the assembler will attempt to use the smallest length possible by
default.

Signed-off-by: Connor McAdams <conmanx360@gmail.com>